Split DTOs into two separate files and renamed fields/classes for increased clarity
